net: remove request throttling
Request throttling is introduced to tackle OOM prolems when requests were very
easily accumulated because of very limited worker threads. Now with short
threads construct, we will process any request ASAP and requests won't occupy
as many memory as before in a short period. So it's time to remove this
throttling because they are never reached in reality.